requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.3
python-docx>=1.1.0